Убрать дублирующие методы
